Kramv - use locals to avoid startup crash on Rosetta2, switch from int to int32_t in kramv, base C++ code for kramv. fix AppIcon
int is 4 bytes on macOS/iOS, but larger on Win.  So replace use with int32_t. This is used in the Kram code, but will be replaced in another update.
M1 start on Rosetta doesn't appear to like static globals and dies.  So eliminate these.  Only two globals.
Split off C++ code from ObjC++ for a Win port.  Most of the Metal/AppKit code isn't portable.