KTX2 - added kram and kramv import support.
KTX2 with/without zstd supercompression files are converted to a KTX1 style KTXImage, so then that can be passed through to the rest of the code.  This requires decompressing all the compressed levels into memory.   A better loader would be to tie this to staging textures and a BlitEncoder to handle uploads.

This viewer can now import and decode if needed.  Works for the test cases.

Must have ktx2ktx2 and ktxsc in the path, and there's a --ktx2 option to kramTextures.py script that uses these to convert ktx to ktx2 files.

Still not supporting KTX2 with BasisLZ compression.
Bug - ETC2 files converted to KTX2 have a missing bottom right pixel in each mip.  The KTX files don't have this.  Tried with and without supercompression, and the missing pixel is still there.