Kram - add version derived from the git tag, and a dummy file that gets replaced in first part of the build script, credits in custom About panel.
cleanup the usage messages, and have them run if "kram encode" for example is run.
add back the (DISABLED) messages when a given encoder is disabled.
Add the simplified format name to info.
Isolate gap value to showSettings so it can be used in multiple places.
Remove version from the cmake files.  This will come from KramVersion.h
Version of kramv and kram are now tied together.  This will help identify if an update is needed.

Ideas for iterating on block size and format that's not quite done.  Would help artists pick an encode format.