kram - add prezero to see if that helps with texture creation.
This is meant to use with shaders that do premul post sampling.  The issue is that if you feed premul to these shaders, then they apply alpha twice.  But on zero pixel areas, these throw off the encoders if the rgb isn't also zero since they weight towards dilation or unseen rgb.