kramv - fix the "Open" and "OpenRecent" menu to work, allow ktx2 support
The UTI's needed to point to an NSDocument for "Open" to not be disabled.

Define a document, and then re-route the load to the view or else the call crashes on readFromData unimplemented in the default NSDocument.

Also define an Export Type Identifier or KTX2 was not highlighted in the NSOpenPanel file listings.

"New" is still now enabled, since I set the app up as an Editor for "KTX2" type.  If they're all viewers, then no "New" option is provided.  Just want that to provide empty document windows.