kram/v - add SUPPORT_RGB flag, fix RGBA32F format
RGB isn't supported by Metal, but we try our best to convert one mip level so that kramv can display the image.  kram converts it to RGBA, but info will still report the RGB format.  Row Alignment isn't always conducive to opening these, but will fix that later.  Also not handling srgb properly.  This has to first go to KramImage to load as RGBA and then back to KTXImage.

Fix RGBA32F format in the table.