kram - break off decoder/encoder from Image, add block decode, add macOS thumbnailer
This is the code to generate a Quicklook thumbnailer.  The project was setup by Xcode, so I don't have it tied to CMake yet.
  Had to move all projects to 10.15, where this frameworks is available.
  This is an app extension for thumbnailing tied into kramv.  Still need to detail the ktx/ktx2 formats in the plist.

Broke out the Encoder/Decoder since they shouldn't be so tied to the single-level Image class.
The thumbnailer wanted to extract a single mip, and render it to CG.  So that's done now.
Simplfied getting mip dimensions, since with mipDown it's a shift and max(1, w).