kramv - fixup the 3d shapes as much as possible, fix eyedropper.
These are still all smushed, since xy-only scale is applied to the box to approximate a wxh image.
Flip the u direction of the sphere/cylinder primitives, since these are inverted from the box.
Invert the bitangent sign, since even with flipping u direction, the tangents are inverted from what they should be.
Split out 2d from 3d view matrices.
Fix eyedropper on archive by setting decodedFormat on that path, and fixing toSnorm8 call use.
Try to activate specular, but it looks bad in ortho and with the scaling.