kram - added KTXImageData to isolate the memory backing a KTXImage
There were too many steps to wrap mmap, file helper, and alias data from a bundle.
Switch FileHelper to size_t.  Used to be int, but that's limited to 32-bit.
Simplify kram setup code for dealing with files.  PNG is still not tied into KTXImage so has to replicate some things.
Bubbling KTXImage up and out of loading.  It's useful to decode blocks and reference in the viewer.