Kramv - simplify loader to only use blit path, also kramv can load png into one level of KTXImage
This lost the mip support of BlitEncoder, but can add that back later or use Mipper.
This opens up pulling PNG from folder, not just KTX/2 files.  Need to call openPNG vs. open call, or detect off 4-5 bytes code at beginning of file.
Remove the cpu upload path.  This is slow and untwiddled, but useful for reference or when synchronous upload required.
Nothing is keeping the KTXImage data alive if the staging buffer is flooded.