kramv - fix sphere prim, add mirrored sphere, disable tan from normal
The author of tan-from-normal removed my inquiry about the algorithm and mirrored uv, so I implemented a mirrored uv test case with a sphere.  Note that the triangles aren't mirrored, only the uvs.  So the normals are flipped, but not the triangles.  Will add that other case later.

Tangent generation can handle that case, since it can compute correct normal/tangent and flip them.  But the tan-from-normal may need the faces inverted as well (and the normals flipped).

Had to rotate the sphere to match the cube/capsule.  It needed pos and normals rotated.  Now lighting is consistent.
Turned off specular in shader for now.
Attenuate diffuse off dot(vertNormal, bumpNormal).  So dark side doesn't have diffuse.