kramv - add shader hotloading
Run buildShaders.sh to rebuild the shaders into a .air file, then to a .metalllib.  These are written to the bin folder, but can be placed anywhere.  Drop this metallib onto the app, and then after running buildShaders.sh, can just select from the recently loaded menu item to hotload.  All shaders and pipelines are rebuilt.  The app starts off using the default.metallib that Xcode bundles into the app.  Can iterate more rapidly on look and feel in kramv this way.