kram - reroute lodepng decompress to miniz
This required some reshuffling of how defaults and decompress are handled in lodepng.  lodepng needed to allocate memory, and also use the defaults that are set. Rich was doing this, and miniz can ignore crc and has faster than 1-bit decode.  I didn't fully update the encode path yet, since it's chopped out.

Add defines to top of lodepng to chop out a bunch of source.  Will eventually move to config.