kramv - add xxx- to supported preview extension names, go back to using thumbnail for the preview instead
There's so little documentation on how to do preview extension on macOS.  This is dead-simple code that loads an image into a CGImage then to the NSImage of an NSImageView.  Setting a red background on the NSImageView shows up 30 to 50% of the time, and the rest of the time nothing shows up.  I've used qlManage to make sure that the end of the code is reached since it will display text from error/print messages, and cavmeman debugging shows the numbers all check out.  The Quick Look Simulator has no docs, and is worthless when hitting run from what I can tell.   It does nothing, not even searching the drive to display a preview.

Add a ViewController to keep QuickLook happy.