kram/v - restore override of srgb state
These blocks are often not set on png files, so need to derive from the -a/-d suffix on the end of the textures.   PS2022 finally added an option to add iccp/gama blocks that reflect sRGB, but most of the test textures don't have this.   Adobe PS and Apple Preview all do srgb * alpha, which is non-linear * linear.  Kram uses linearToSrgb( srgbToLinear( srgb ) * alpha ), which is then linear color and linear alpha.  This means some alpha and glows look different in kramv.  Use isPremulRGB to approximate Photoshop/Preview which doesn't set an sRGB format on the image.

Eventually will make this override optional, but don't have time to expose that.