kramv - fix pinch-zoom and display of non-square images
All images used to be tied to a square, and so non-uniform scaling was used to match non-square images.  But that code was modified to create a quad with the correct aspect ratio instead.  But this threw off pinch-zoom lock to the cursor, eyedropper, and correct aspect ratio display.

Code now uses uniform scaling even in 2D mode, and then adjusts the pinch-zoom calcs.