kram - turn off FASTL for now
The map/set and unordered_map/set behavior likely isn't shippable.  It shifts the address of the values because they are insertion sorted into an array.  All that copying is costly too.  This is likely why a linked-list is used by std, but that has poor allocation locality and allocates one value at a time unless a chunk allocator is specified.