kramv - add back srgb toggle (key 9) for png
The problem is that encoded formats and mips already have srgb baked in.  So currently this only applies to PNG, where kramv may have selected the wrong srgb setting if the name doesn't identify the content type clearly.  The fallback is to turn on srgb to match other tools.

This uses a MTLTextureView to disable srgb, so no more memory is needed.  Also the usage view setting doesn't have to be set for just a format change.

Can see where this does work on TestColorGradient.png and doesn't work on TestAlphaGradient.png.  This is because the all white color alpha gradient is the same whether in srgb or not.  But when dealing with premul, the alpha has modded the color itself, and will be affected by the srgb/non-srgb of the format.