kram - for now turn off using the srgb state of srcImage
This was causing files from figma which are linear color to do srgb -> lin conversion.  This caused mismatches with files converted to ktx from apps like Photoshop which don't default to setting srgb on png files unles a checkbox is clicked.  Note that neither app actually mods the pixels.  127 = 127 output. Now the conversion only occurs if -srgb (dst) is set which was already happening in initWithArgs.