kram - add mipflood, remove in-place mips, fix timer
Add mipflooding.  This build the entire mip chain, and them upsamples it to the higher mip.  This is a cheaper dilate, and also compresses better, since many of the regions have blocky pixels.  Still unclear if using premul is better than doing this.  But it works for God of War.

So this has to allocate more memory for the smaller mip chain, since it no longer does in-place mips.  The smaller mips are only 1/3rd of the largest mip.  Skip mips in place, and then build out the rest of the mip chain.  This allows

Fix the timer by replacing the clock with system specific calls.  The clock::now() calls seem to have a granularity of 32ms which is terrible.   Go back to system calls that I know are in nanoseconds.  Was also getting an assert by initializing startTime within the call with clock().  That would return a time delta of 0, so init those things in static init.