kram - hlsl code cleanup, make work with glslc
glslc and dxc are competing compilers for HLSL in the Vulkan SDK.  Make sure code works with either.
Simplify the code to mostly a single array of type data.  This avoids a ton of code switch statements that can get out of sync.
Update readme with more info.  It needs more work.