kram - update buildShaders.sh
Get this to gen reflection data, but it seems to be in a myriad of formats.   At least with spriv transpiles, can spriv-reflect the spv files themselves.
Need to show a comparison of transpiled MSL vs. what hlslparser generates.

TODO: get // comments preserved and over to the .metal/.hlsl file
TODO: Syntax highlighting not working in Xcode for HLSL or for most of the generated .metal file.  Xcode's the worst for syntax highlighting of any file they don't want to deal with.   VSCode does quite nicely with HLSL highlighting.