kram - hlslparser update
Use brew ninja build script for now.   It's fast.  Don't know why Xcode uses it's own sanitized path, instead of my own with brew and vulkan sdk in it.  Turn on -line to jump back to errors in original source files from IDE.  Turn on warnings as errors, just to keep the code clean.   This makes output hard to read, so recommend disabling it when developing the conversions.  VS supports same gcc/clang error/warning report, so use that.