kram - update hlslparser
Bump to SM2.4, since this seems to be first support for separate metallibsym.  Was getting warnings about it being unsupported, but the warnings were far from clear, and only generated periodically.

Move the headers back to shaders directory, since can set -I include path, and jump to original errors with the -line directive.  That's super nice.