kram - remove -D_HAS_EXCEPTIONS=0
Advice is all over the place, but this is supposed to disable exceptions only within STL.

See here about letting VC set this.  The code was all compiling when CMakeLists.txt was setting this flag incorrectly. But setting it gens a ton of errors and warnings across the Win code base.  Check tests for _HAS_EXCEPTIONS in fmt/core.h that suddenly doesn't compile, etc.

https://reviews.llvm.org/D52998