kram - replace simdjson with json11, a blocked linear allocator, and immutable strings
This is a fairly efficient json reader/writer.  I've taken the now defunct json11 packed from Dropbox, and heavily modified it.  So keys and strings are aliased.  And there are almost not allocation during reading.  Still more work to do on writer just in helping build up the json.  This also supports iterators.   One slow part is addJson which must search to find the end of the linked list.  But all allocations are block allocated in 1024 elements, DOM nodes are 32B currently, and a 100KB atlas file takes about 600KB to read.  And that's all in nice 32KB chunks from the heap.  This still needs to be fixed to handle trailing comma.

No guarantees this is the fastet json parser, and I'll likely bring back simdjson.  But I like how simple my json11 parser is.

Added a BlockedLinearAllocator.  Just wanted something that could produce a single node, be reset, and then resupply those nodes.  There is an item count per block, and item size.  No variable block support.  But this works well for Json nodes which are all 32B.

Comment out FASTL.  This stl just doesn't work out.
Turn off stack traces for errors for now.

Added an ImmutableString/Pool.  I always use this class, and so it's time to add it in

Fixed crash where atlas wasn't getting cleared out when loading a different file, and the hud would try to display from a dead lastAtlas entry.