kram-profile - a tool to look at profile data
Clang, krams' VMA python converter, and games can generate Catapult json traces that also load into Perfetto.  This is a SwiftUI-based app that displays a file or folder, and then opens selected files into a list navigation view.   That way can quickly review the traces.

The Perfetto window loads the first {"perfetto":{}} json packet sent over from Swift to the WKWebView, but then stops.  Something seems to be latching the state, and I really don't want to have to reload() the page and cause the UI to flash.  Find out why this is happening.

More work to tie this in with CBA, and also fix up clang json files created using -ftime-trace.