kram-profile - add gzip support, add reload
No duration on this for now on gzip, but feed direct to Perfetto and it can ungzip it.  CBA won't read gzip though.   Also clang json files still need fixed up.  Perfetto is working on zip support.  May need to export KramZipHelper as a clang module, or make a framework.