kram-profile - deal with more stupid SwiftUI bugs
Try to create a sorted and searchable List in SwiftUI, and then pull your hair out.
If add .searchable, then the editText focuses itself and can't tab-switch focus anywhere else.  This totally blocks any interaction with the List or WebView.  This some horrible bug in SwiftUI, but it's hard to tell with so much secret code changing state in the black box.
Turn on Swift C++ support