kram-profile - more archive support, ignore unsupported files
Notes on archives which are tricky due to storage of relative paths.  These can confuse the List id mechanism which is based on URL.  Just make sure to store unique identifying folder when building archives.
Can now build a report based on selected file - this will look at archive or same parent folder.  This allows multiple archives to be loaded.