kram - add ZipStream and Perf
These are a gzip stream that can compress data passed to it.  Perfetto can take in gz and zip compressed files.  Gzip has a 10B header + 8B footer, and this uses miniz deflate calls.  Had to disable the default zlib header/footer, but would be interesting to support that too.  Use this to unpack and keep the original file, since this is compress only right now.   gzunzip -k file.gz

Needed to expose some of the thread calls for Perf to obtain the id and name.