kram - turn off App Sandbox in entitlements
This basically sets com.apple.quarantine on any files written or read by an app.  This is problematic, since trying to write profile from kramv, open that, and have it show up in the profiler.  Will revisit and assign runtime hardening, sandbox, and other items to only the app store build.  It's confusing since Xcode also has an "App Sandbox" setting that is set to No.