kram - move to xcconfig
Was tired of changing settings across the projects, so it's time.   Also x64 needs settings that then cause the arm64 to log warnings.

For some reason Xcode UI keeps copying the kram.config file when set on GLTF or GLTFMTL instead of referencing the common one in build2.  So I had to symlink to those.