kram - update cibiuld.sh to put libkram, libvectormath into dirs
Since this is all on mutliplatform build, the only way to uniquify the libs is to either put them in a folder, or packages those folders into an xcframework
So this is the first part of putting them in folders.   Previously the target would give them a -ios or -vos suffis to the library name.