kram - fix odd mipgen, fix cube/array/cubearray texturegen
48x8 image was generating incorrect top mips.  Needed to handle special 1x1 and 3x1 case on the smallest mips.
The mipgen was clobbering the srcImage.pixels fields, so restore that when processing a new chunk.
bc7enc change - assign error to alpha when it tries mode 6 on an opaque texture.  Otherwise, the alpha error is not accounted for.  Black255 -> Black254.  So I added a Black4x4-a.png test case, and the b+w checkerboard already demonstrates this.

Attempt to fix this issue in the old .cpp encoder.  Haven't tried the ispc, since that is harder to integrate and a big dependency.
https://github.com/richgel999/bc7enc_rdo/issues/15

Turn off simd config printout.